Query Nuxt Content files without trailing slash
Using the queryContent composable with a path containing a trailing
slashes returns a 404 because Nuxt Content creates internal paths for
Content files without a trailing slash (see
https://content.nuxtjs.org/guide/writing/content-directory#paths).

We need to support routes with trailing slashes because the Nuxt
prerendering (SSG) creates a directory containing an index.html file for
each content file e.g. test.md -> test/index.html. Github Pages
recognises that it is accessing a directory and will always append a
trailing slash when a route is accessed.

This behaviour is not configurable so we need to strip any trailing
slash from the pathname of the URL before querying Nuxt Content.

Test plan:
- Navigate to route managed by Nuxt Content and confirm that refreshing
  the page works after Github Pages appends a trailing slash