🔧 Update build script to support guides directory from both jekyll and nuxt site
We have existing pages docs.centrapay.com/guides/* living on the jekyll site.
For the nuxt site we want the URL docs.centrapay.com/guides/farmlands-pos-integration.
Update the build script so that the guides directory can be merged.

Test plan: Expect to be able to navigate to existing /guides and /guides/farmlands-pos-integration