FL partner page feedback adjustments
The main image has been resized so it doesn't take up so much room. Also added a nuxt hook to instantly scroll to top on page navigation. This allows us to still have smooth-scrolling on our nuxt content pages. There may be a better solution for this but this works for now