Fix bug where direct navigation to Farmlands POS guide page keeps nav menu closed
The problem is to do with trailing slashes being added by GitHub pages
when refreshing the page or directly navigating to a page.

We should change where the site is hosted as it seems like this
behaviour is not configurable in GitHub pages.

There is a small UI bug where after reloading or directly navigating to
the Farmlands POS guide page, opening and closing the nav menu
un-highlights the Farmlands POS guide in the nav menu. This should be
fixed after changing the site host.