⏮️ Revert: Keep Nuxt default scroll behaviour
A hook had been added to scroll to the top instantly when on a new page.
This was because the following behaviour was undesirable: Scrolling to bottom of Farmlands Partner page and clicking 'POS Integration Guide' resulted in the POS Integration Guide loading and then scrolling automatically from the bottom to the top of the page.
However, this caused a regression as if your URL led to a title, e.g. https://docs.centrapay.com/guides/farmlands-pos-integration/#pre-auth-flow you would end up at the top of the page.