🐛 Fix: Add leading slash to active nav map
Not having the leading slash means that if you are on a page, then refresh the browser, the navigation bar would not show the active URL as intended.

Test plan: View POS guide, refresh page and expect Reference Nav to be open and Point of Sale to have a background, indicating it is active.