Save disclosure state when routing between pages 💾
When routing between pages all Vue state resets. This was causing the site navigation disclosures to close every time you changed a page, apart from the disclosure that has a link to your current page.
Change to save open disclosures in local storage. This state is read when creating the disclosure so it can determine whether the disclosure is open.