Remove markdoc
We have decided not to move forward with Markdoc at this stage as there are a few issues that are preventing us from delivering the main objective - to migrate the legacy Jekyll docs to the new Astro site. Once we have migrated the site we can look into solving some of the pain points (e.g. code block styling) before converting the mdx into Markdoc