Update UaaS Documentation (#960)
The existing documentation indicated that integrators could throw whatever errors they wanted from each endpoint. This is not how we want people to integrate. Errors should only be thrown if something unexpected is occurring in their system. We expect 2XX responses for successful and failed transactions. All integrators must implement the GET transaction endpoint whether or not they process transactions asynchronously. This allows us to resolve transactions in future if we receive errors.