Update Get Managed Integration (#1004)
We have reworked this endpoint to retrieve the managed integration by type and external id instead of by the integration id. This means third parties do not need to store our integration id, they only keep a record of the external id (The unique identifier they are already keeping track of).